requirements.txt
argparse
requests
numpy
tsp_solver2
